lab9/authentication-lab/requirements.txt
autopep8==1.4.3
certifi==2018.11.29
chardet==3.0.4
Django==2.1.7
django-filter==2.1.0
djangorestframework==3.9.1
httpie==1.0.2
idna==2.8
Markdown==3.0.1
pycodestyle==2.5.0
Pygments==2.3.1
pytz==2018.9
requests==2.21.0
urllib3==1.24.1
